backend/requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.1
google-genai==0.6.0
httpx==0.27.2
pdfplumber==0.11.9
pydantic==2.12.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dotenv==1.2.2
reportlab==4.2.5
structlog==25.5.0
supabase==2.10.0
uvicorn==0.41.0
